mula

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,7 +6,7 @@
  */
 int main(void)
 {
-	int m, l, n, fib;
+	long int m, l, n, fib;
 	
 	l = 1;
 	m = 2;
